simulation/requirements.txt
kafka-python>=2.0.2
lz4>=4.0.0
neo4j>=5.18.0
qdrant-client>=1.9.0,<1.10.0
psycopg2-binary>=2.9.9
requests>=2.31.0
hvac>=2.1.0
ff3>=1.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
pyyaml>=6.0
